extracts models and controllers to seperate files for each endpoint, reflecting proper MVC
