+++ api/controllers/app.api.controllers.js
@@ -1,0 +1,9 @@
+const {fetchEndpoints} = require('../models/app.api.models')
+
+exports.getEndpoints = (req, res, next) => {
+  fetchEndpoints()
+    .then((endpoints) => {
+      res.status(200).send(endpoints);
+    })
+    .catch(next);
+};
