+++ api/controllers/app.api.topics.controllers.js
@@ -1,0 +1,13 @@
+const { fetchTopics } = require("../models/app.api.topics.models");
+
+exports.getTopics = (req, res, next) => {
+  fetchTopics()
+    .then((topics) => {
+      res.status(200).send({ topics });
+    })
+    .catch(next);
+};
+
+
+
+
