--- api/controllers/app.topics.controllers.js
@@ -1,17 +1,0 @@
-const { fetchTopics, fetchEndpoints } = require("../models/app.topics.models");
-
-exports.getTopics = (req, res, next) => {
-  fetchTopics()
-    .then((topics) => {
-      res.status(200).send({ topics });
-    })
-    .catch(next);
-};
-
-exports.getEndpoints = (req, res, next) => {
-  fetchEndpoints()
-    .then((endpoints) => {
-      res.status(200).send(endpoints);
-    })
-    .catch(next);
-};
