+++ api/models/app.api.topics.models.js
@@ -1,0 +1,8 @@
+const db = require("../../db/connection");
+
+exports.fetchTopics = () => {
+  const sqlQuery = "SELECT * FROM topics";
+  return db.query(sqlQuery).then((topics) => topics.rows);
+};
+
+
