+++ db/query-test-db.sql
@@ -1,0 +1,3 @@
+\c nc_news_test
+
+SELECT * FROM articles
