+++ package.json
@@ -7,7 +7,8 @@
     "setup-dbs": "psql -f ./db/setup.sql",
     "seed": "node ./db/seeds/run-seed.js",
     "test": "jest",
-    "prepare": "husky install"
+    "prepare": "husky install",
+    "query-test-db": "psql -f ./db/query-test-db.sql > ./db/test-db-queries.txt"
   },
   "repository": {
     "type": "git",
